+++ src/main/java/org/example/advaned/leehyemin/unitthree/MusicPlayable.java
@@ -1,0 +1,10 @@
+package org.example.advaned.leehyemin.unitthree;
+
+public interface MusicPlayable {
+    public void play();
+    public void pause();
+    public void stop();
+    public void nextSong();
+    public void previousSong();
+
+}
