+++ src/main/java/org/example/advaned/leehyemin/unitthree/PowerControllable.java
@@ -1,0 +1,6 @@
+package org.example.advaned.leehyemin.unitthree;
+
+public interface PowerControllable {
+    public void powerOn();
+    public void powerOff();
+}
