+++ src/main/java/org/example/advaned/leehyemin/unitthree/SmartDevice.java
@@ -1,0 +1,9 @@
+package org.example.advaned.leehyemin.unitthree;
+
+public interface SmartDevice {
+    void turnOn();
+    void turnOff();
+    void setTemperature(int temp);
+    void setTimer(int seconds);
+    void playMusic(String song);
+}
