+++ src/main/java/org/example/advaned/leehyemin/unitthree/TempControl.java
@@ -1,0 +1,8 @@
+package org.example.advaned.leehyemin.unitthree;
+
+public class TempControl implements TempControllable{
+    @Override
+    public void setTemperature(int temp) {
+        System.out.println(temp+"도로 온도 변경");
+    }
+}
