+++ src/main/java/org/example/advaned/leehyemin/unitthree/TempControllable.java
@@ -1,0 +1,5 @@
+package org.example.advaned.leehyemin.unitthree;
+
+public interface TempControllable {
+    public void setTemperature(int temp);
+}
