+++ src/main/java/org/example/advaned/leehyemin/unitthree/TimerControllable.java
@@ -1,0 +1,5 @@
+package org.example.advaned.leehyemin.unitthree;
+
+public interface TimerControllable {
+    public void timerControl(int time);
+}
