+++ src/main/java/org/example/advaned/leehyemin/unitthree/VolumeControl.java
@@ -1,0 +1,14 @@
+package org.example.advaned.leehyemin.unitthree;
+
+public class VolumeControl implements VolumeControllable {
+
+    @Override
+    public void volumeUp() {
+        System.out.println("Volume Up");
+    }
+
+    @Override
+    public void volumeDown() {
+        System.out.println("Volume Down");
+    }
+}
