+++ src/main/java/org/example/advaned/leehyemin/unitthree/VolumeControllable.java
@@ -1,0 +1,6 @@
+package org.example.advaned.leehyemin.unitthree;
+
+public interface VolumeControllable {
+    public void volumeUp();
+    public void volumeDown();
+}
