+++ src/main/java/org/example/advaned/leehyemin/unittwo/Notification.java
@@ -1,0 +1,5 @@
+package org.example.advaned.leehyemin.unittwo;
+
+public interface Notification {
+    public void send();
+}
