--- src/main/java/org/example/basic/leehyemin/unitfour/SmartHome/MusicPlayable.java
@@ -1,10 +1,0 @@
-package org.example.basic.leehyemin.unitfour.SmartHome;
-
-public interface MusicPlayable {
-    public void play();
-    public void pause();
-    public void stop();
-    public void nextSong();
-    public void previousSong();
-
-}
