--- src/main/java/org/example/basic/leehyemin/unitfour/SmartHome/PowerControl.java
@@ -1,13 +1,0 @@
-package org.example.basic.leehyemin.unitfour.SmartHome;
-
-public class PowerControl implements PowerControllable{
-    @Override
-    public void powerOn() {
-        System.out.println("전원이 켜졌습니다.");
-    }
-
-    @Override
-    public void powerOff() {
-        System.out.println("전원이 꺼졌습니다.");
-    }
-}
