--- src/main/java/org/example/basic/leehyemin/unitfour/SmartHome/PowerControllable.java
@@ -1,6 +1,0 @@
-package org.example.basic.leehyemin.unitfour.SmartHome;
-
-public interface PowerControllable {
-    public void powerOn();
-    public void powerOff();
-}
