--- src/main/java/org/example/basic/leehyemin/unitfour/SmartHome/TempControl.java
@@ -1,8 +1,0 @@
-package org.example.basic.leehyemin.unitfour.SmartHome;
-
-public class TempControl implements TempControllable{
-    @Override
-    public void setTemperature(int temp) {
-        System.out.println(temp+"도로 온도 변경");
-    }
-}
