--- src/main/java/org/example/basic/leehyemin/unitfour/SmartHome/TempControllable.java
@@ -1,5 +1,0 @@
-package org.example.basic.leehyemin.unitfour.SmartHome;
-
-public interface TempControllable {
-    public void setTemperature(int temp);
-}
