--- src/main/java/org/example/basic/leehyemin/unitfour/SmartHome/TimeControl.java
@@ -1,8 +1,0 @@
-package org.example.basic.leehyemin.unitfour.SmartHome;
-
-public class TimeControl implements TimerControllable {
-    @Override
-    public void timerControl(int time) {
-        System.out.println(time+"시간 후에 종료됩니다.");
-    }
-}
