--- src/main/java/org/example/basic/leehyemin/unitfour/SmartHome/TimerControllable.java
@@ -1,5 +1,0 @@
-package org.example.basic.leehyemin.unitfour.SmartHome;
-
-public interface TimerControllable {
-    public void timerControl(int time);
-}
