--- src/main/java/org/example/basic/leehyemin/unitfour/SmartHome/VolumeControl.java
@@ -1,14 +1,0 @@
-package org.example.basic.leehyemin.unitfour.SmartHome;
-
-public class VolumeControl implements VolumeControllable {
-
-    @Override
-    public void volumeUp() {
-        System.out.println("Volume Up");
-    }
-
-    @Override
-    public void volumeDown() {
-        System.out.println("Volume Down");
-    }
-}
