--- src/main/java/org/example/basic/leehyemin/unitfour/SmartHome/VolumeControllable.java
@@ -1,6 +1,0 @@
-package org.example.basic.leehyemin.unitfour.SmartHome;
-
-public interface VolumeControllable {
-    public void volumeUp();
-    public void volumeDown();
-}
