--- src/main/java/org/example/basic/leehyemin/unitfour/Student/Student.java
@@ -1,27 +1,0 @@
-package org.example.basic.leehyemin.unitfour.Student;
-
-public class Student {
-    String name;
-    int score;
-
-    public Student(String name, int score) {
-        this.name = name;
-        this.score = score;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public int getScore() {
-        return score;
-    }
-
-    public void setScore(int score) {
-        this.score = score;
-    }
-}
