--- src/main/java/org/example/basic/leehyemin/unitthree/notification/Email.java
@@ -1,8 +1,0 @@
-package org.example.basic.leehyemin.unitthree.notification;
-
-public class Email implements Notification {
-    @Override
-    public void send() {
-        System.out.println("Email 전송 완료");
-    }
-}
