--- src/main/java/org/example/basic/leehyemin/unitthree/notification/Notification.java
@@ -1,5 +1,0 @@
-package org.example.basic.leehyemin.unitthree.notification;
-
-public interface Notification {
-    public void send();
-}
