--- src/main/java/org/example/basic/leehyemin/unitthree/notification/Push.java
@@ -1,8 +1,0 @@
-package org.example.basic.leehyemin.unitthree.notification;
-
-public class Push implements Notification {
-    @Override
-    public void send() {
-        System.out.println("Push 알림 전송 완료");
-    }
-}
