--- src/main/java/org/example/basic/leehyemin/unitthree/notification/SMS.java
@@ -1,8 +1,0 @@
-package org.example.basic.leehyemin.unitthree.notification;
-
-public class SMS implements Notification {
-    @Override
-    public void send() {
-        System.out.println("SMS 전송 완료");
-    }
-}
