--- src/main/java/org/example/basic/leehyemin/unitthree/shopping/ShoppingPoint.java
@@ -1,20 +1,0 @@
-package org.example.basic.leehyemin.unitthree.shopping;
-
-public class ShoppingPoint {
-    public void earnPoints(int price, Member mall) {
-        String grade=mall.getGrade();
-
-        double rate=0.01;
-
-        if(grade == "SIVER"){
-            rate += 0.01;
-        } else if(grade == "GOLD"){
-            rate += 0.02;
-        }
-
-        int earnPoints = (int) (price * rate);
-        mall.addPoint(earnPoints);
-
-        System.out.println("적립 포인트 : "+mall.getPoint());
-    }
-}
